fix(inventory): rename core -> to dao cmd
